[REVERT ME] Disable kernel compatibility check
Once device is booted up, error dialog "There\'s an internal problem
with your device. Contact  your manufacturer for details." is shown.

Kernel compatibility check fails because of kernel version and
target level incompatibility. As per Android 14, target-level needs to
be set to 8. Since not all vendor HALs are compatible as per
target-level 8, target-level is set to 7 in device manifest.

Disable the kernel compatibility check until the HALs are compatible
with target-level 8 and target-level is changed to 8.

Tracked-On: OAM-111022
Change-Id: I8a8e50e19b1ebe812bb15abf1a45f94d18b7b20b
Signed-off-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>